+++ app/Http/Controllers/SupplyController.php
@@ -317,11 +317,8 @@ class SupplyController extends Controller
         if ('POST' === $request->getRealMethod()) {
             $input = $request->input();
         } elseif ('GET' === $request->getRealMethod()) {
-            if (null === ($query = $request->getQueryString())) {
-                throw new UnprocessableEntityHttpException('Query is required');
-            }
-            if (false !== ($index = strpos($query, '&'))) {
-                $query = substr($query, 0, $index);
+            if (null === ($query = $request->query('data'))) {
+                throw new UnprocessableEntityHttpException('Query `data` is required');
             }
             $input = json_decode(Utils::urlSafeBase64Decode($query), true);
         } else {
@@ -331,46 +328,46 @@ class SupplyController extends Controller
         if (!is_array($input)) {
             throw new UnprocessableEntityHttpException('Invalid body type');
         }
-        foreach (['page', 'placements'] as $field) {
+        foreach (['context', 'placements'] as $field) {
             if (!isset($input[$field])) {
                 throw new UnprocessableEntityHttpException(sprintf('Field `%s` is required', $field));
             }
         }
 
-        if (!is_array($input['page'])) {
-            throw new UnprocessableEntityHttpException('Field `page` must be an object');
+        if (!is_array($input['context'])) {
+            throw new UnprocessableEntityHttpException('Field `context` must be an object');
         }
-        $page = $input['page'];
+        $context = $input['context'];
         foreach (['iid', 'url'] as $field) {
-            if (!isset($page[$field])) {
-                throw new UnprocessableEntityHttpException(sprintf('Field `page.%s` is required', $field));
+            if (!isset($context[$field])) {
+                throw new UnprocessableEntityHttpException(sprintf('Field `context.%s` is required', $field));
             }
-            if (!is_string($page[$field])) {
-                throw new UnprocessableEntityHttpException(sprintf('Field `page.%s` must be a string', $field));
+            if (!is_string($context[$field])) {
+                throw new UnprocessableEntityHttpException(sprintf('Field `context.%s` must be a string', $field));
             }
         }
-        if (array_key_exists('metamask', $input['page'])) {
-            if (!is_bool($input['page']['metamask'])) {
-                throw new UnprocessableEntityHttpException('Field `page.metamask` must be a boolean');
+        if (array_key_exists('metamask', $context)) {
+            if (!is_bool($context['metamask'])) {
+                throw new UnprocessableEntityHttpException('Field `context.metamask` must be a boolean');
             }
         }
         if (!is_array($input['placements'])) {
             throw new UnprocessableEntityHttpException('Field `placements` must be an array');
         }
 
-        $isDynamicFind = array_key_exists('publisher', $page);
+        $isDynamicFind = array_key_exists('publisher', $context);
         if ($isDynamicFind) {
-            if (!is_string($page['publisher'])) {
-                throw new UnprocessableEntityHttpException('Field `page.publisher` must be a string');
+            if (!is_string($context['publisher'])) {
+                throw new UnprocessableEntityHttpException('Field `context.publisher` must be a string');
             }
-            if (!isset($page['medium'])) {
-                throw new UnprocessableEntityHttpException('Field `page.medium` is required');
+            if (!isset($context['medium'])) {
+                throw new UnprocessableEntityHttpException('Field `context.medium` is required');
             }
-            if (!is_string($page['medium'])) {
-                throw new UnprocessableEntityHttpException('Field `page.medium` must be a string');
+            if (!is_string($context['medium'])) {
+                throw new UnprocessableEntityHttpException('Field `context.medium` must be a string');
             }
-            if (isset($page['vendor']) && !is_string($page['vendor'])) {
-                throw new UnprocessableEntityHttpException('Field `page.vendor` must be a string');
+            if (isset($context['vendor']) && !is_string($context['vendor'])) {
+                throw new UnprocessableEntityHttpException('Field `context.vendor` must be a string');
             }
 
             foreach ($input['placements'] as $placement) {
@@ -393,13 +390,13 @@ class SupplyController extends Controller
         if ($isDynamicFind) {
             try {
                 $medium = $configurationRepository->fetchMedium(
-                    $page['medium'],
-                    $page['vendor'] ?? null
+                    $context['medium'],
+                    $context['vendor'] ?? null
                 );
             } catch (InvalidArgumentException $exception) {
                 throw new UnprocessableEntityHttpException($exception->getMessage());
             }
-            $site = $this->getSiteOrFail($page);
+            $site = $this->getSiteOrFail($context);
 
             foreach ($input['placements'] as $key => $placement) {
                 $zoneType = $this->getZoneType($placement);
@@ -430,15 +427,11 @@ class SupplyController extends Controller
             }
         }
 
-        foreach ($input['placements'] as $key => $placement) {
-            $input['placements'][$key]['options'] = [
-                'banner_type' => $placement['types'] ?? null,
-                'banner_mime' => $placement['mimeTypes'] ?? null,
-            ];
-        }
-
-        $foundBanners = $this->findBanners($input, $request, $response, $contextProvider, $bannerFinder);
-        return self::json(['data' => $foundBanners->map($this->mapFoundBannerToResult())]);
+        $mappedInput = self::mapFindInput($input);
+        $foundBanners = $this->findBanners($mappedInput, $request, $response, $contextProvider, $bannerFinder)
+            ->map($this->mapFoundBannerToResult())
+            ->toArray();
+        return self::json(['data' => $foundBanners]);
     }
 
     public function findSimple(
@@ -1048,7 +1041,7 @@ class SupplyController extends Controller
 
         $userId = User::fetchByUuid($case->publisher_id)->id;
 
-        Storage::disk('local')->put(
+        Storage::disk('local')->append(
             'reported-ads.txt',
             sprintf('%s;%s', $userId, $bannerId)
         );
@@ -1127,7 +1120,9 @@ class SupplyController extends Controller
             try {
                 $payoutAddress = WalletAddress::fromString($publisher);
             } catch (InvalidArgumentException) {
-                throw new UnprocessableEntityHttpException('Field `page.publisher` must be an ID or account address');
+                throw new UnprocessableEntityHttpException(
+                    'Field `context.publisher` must be an ID or account address'
+                );
             }
             $user = User::fetchByWalletAddress($payoutAddress);
             if (null === $user) {
@@ -1153,16 +1148,16 @@ class SupplyController extends Controller
         return $user;
     }
 
-    private function getSiteOrFail(array $page): Site
+    private function getSiteOrFail(array $context): Site
     {
-        $user = $this->getPublisherOrFail($page['publisher']);
+        $user = $this->getPublisherOrFail($context['publisher']);
 
         try {
             $site = Site::fetchOrCreate(
                 $user->id,
-                $page['url'],
-                $page['medium'],
-                $page['vendor'] ?? null,
+                $context['url'],
+                $context['medium'],
+                $context['vendor'] ?? null,
             );
         } catch (InvalidArgumentException $exception) {
             throw new UnprocessableEntityHttpException($exception->getMessage());
@@ -1191,6 +1186,33 @@ class SupplyController extends Controller
             $zoneType = null;
         }
         return $zoneType;
+    }
+
+    private static function mapFindInput(array $input): array
+    {
+        $mapped = [
+            'page' => [
+                'iid' => $input['context']['iid'],
+                'url' => $input['context']['url'],
+                'metamask' => (int)($input['context']['metamask'] ?? 0),
+            ],
+        ];
+
+        foreach ($input['placements'] as $placement) {
+            $placementData = [
+                'placementId' => $placement['placementId'],
+                'options' => [
+                    'banner_type' => $placement['types'] ?? null,
+                    'banner_mime' => $placement['mimes'] ?? null,
+                ],
+            ];
+            if (isset($placement['id'])) {
+                $placementData['id'] = $placement['id'];
+            }
+            $mapped['placements'][] = $placementData;
+        }
+
+        return $mapped;
     }
 
     private function mapFoundBannerToResult(): Closure
@@ -1222,7 +1244,7 @@ class SupplyController extends Controller
     {
         $fieldsOptional = [
             'types',
-            'mimeTypes',
+            'mimes',
         ];
         foreach ($fieldsOptional as $field) {
             if (isset($placement[$field])) {