+++ tests/app/Http/Controllers/Manager/ApiCampaignsControllerTest.php
@@ -50,7 +50,7 @@ final class ApiCampaignsControllerTest extends TestCase
         'createdAt',
         'updatedAt',
         'type',
-        'mimeType',
+        'mime',
         'hash',
         'scope',
         'name',